Add feature in which user entered data remains in the form if the data validation fails.
Uses if control statements in the view to determine
if there is a failed event and if the failed event has
a field (name, title, etc.). If the field exists then
on the new get request it is populated with the prior
input.

Just thought of flash hash and may research...